Reestructuración del proyecto
- Se ha reemplazado la carpeta 'app' por `core`.
- Se han actualizado las pantallas y la navegación existentes para utilizar las nuevas rutas y servicios.
- Se han refactorizado los formularios de inicio de sesión y los componentes de inicio de sesión del administrador para mejorar la organización.
- Se han eliminado las funciones obsoletas y se han limpiado las importaciones en todo el proyecto.
