--- lib/app/core/core.dart
@@ -1,8 +1,0 @@
-// Constants
-export '../constants/constants.dart';
-
-// Routes
-export '../routes/routes.dart';
-
-// Data Services
-export '../data/data.dart';
