--- lib/app/data/data.dart
@@ -1,2 +1,0 @@
-// Auth Service
-export 'auth_service.dart';
