+++ lib/core/core.dart
@@ -1,0 +1,8 @@
+// Constants
+export 'constants/constants.dart';
+
+// Routes
+export 'routes/routes.dart';
+
+// Data Services
+export 'data/data.dart';
