+++ lib/features/auth/screens/login_reclutador_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../widgets/header.dart';
-import '../widgets/login_form.dart';
+import 'package:flutter_app/features/auth/widgets/login_form.dart';
+import 'package:flutter_app/features/shared/widgets/header.dart';
 
 class LoginReclutadorScreen extends StatelessWidget {
   const LoginReclutadorScreen({super.key});
@@ -19,12 +19,12 @@ class LoginReclutadorScreen extends StatelessWidget {
           ],
         ),
       ),
-      resizeToAvoidBottomInset: true, 
-      body: SingleChildScrollView(  
+      resizeToAvoidBottomInset: true,
+      body: SingleChildScrollView(
         child: Column(
           children: <Widget>[
             SizedBox(height: 20),
-            LoginForm(0),  // El formulario de login
+            LoginForm(0), // El formulario de login
             SizedBox(height: 20),
             Padding(
               padding: const EdgeInsets.all(20.0),
