+++ lib/features/auth/widgets/login_form.dart
@@ -1,115 +1,143 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/app/core/core.dart';
+import 'package:flutter_app/features/auth/services/user_services.dart';
+import 'package:flutter_app/features/recruiter/main_reclutadores.dart';
+import 'package:flutter_app/features/student/main_estudiantes.dart';
 
 class LoginForm extends StatefulWidget {
-  const LoginForm({super.key});
+  final int index;
+  const LoginForm(this.index, {super.key});
 
   @override
-  LoginFormState createState() => LoginFormState();
+  State<StatefulWidget> createState() => _LoginFormState();
 }
 
-class LoginFormState extends State<LoginForm> {
+class _LoginFormState extends State<LoginForm> {
   final _formKey = GlobalKey<FormState>();
-  String _email = '';
-  String _password = '';
+  late TextEditingController _email;
+  late TextEditingController _password;
 
   void _submit() async {
     if (_formKey.currentState!.validate()) {
-      _formKey.currentState!.save();
+      final response = await UserServices.login(_email.text, _password.text,
+          widget.index == 0 ? 'Reclutador' : 'Estudiante');
+      if (!mounted) return;
 
-      // Guardar el estado de login
-      await AuthService.login(userType: 'admin');
-
-      // Navegar al home screen
-      if (mounted) {
-        Navigator.pushReplacementNamed(context, AppRoutes.admin);
+      if (response.containsKey('token')) {
+        UserServices.setToken(response['token']!);
+        _formKey.currentState!.save();
+        print('Email: $_email, Password: $_password');
+        Navigator.pushAndRemoveUntil(
+          context,
+          MaterialPageRoute(
+            builder: (context) =>
+                widget.index == 0 ? MainReclutadores() : MainEstudiantes(),
+          ),
+          (Route<dynamic> route) => false,
+        );
+      } else {
+        showDialog(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: Text('Error'),
+              content: Text(
+                  'Error de inicio de sesión. Por favor, verifique sus credenciales.'),
+              actions: <Widget>[
+                TextButton(
+                  child: Text('OK'),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  },
+                ),
+              ],
+            );
+          },
+        );
       }
     }
   }
 
   @override
+  void initState() {
+    _email = TextEditingController(text: "");
+    _password = TextEditingController(text: "");
+    super.initState();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      child: Column(
-        children: <Widget>[
-          SizedBox(
-            width: 275.39,
-            height: 75,
-            child: TextFormField(
-              decoration: InputDecoration(
-                labelText: 'Código admin',
-                filled: true,
-                fillColor: Color(0xFFDBE2F6),
-                enabledBorder: OutlineInputBorder(
-                  borderSide: BorderSide(color: Color(0xFF1E3984), width: 1.0),
-                  borderRadius: BorderRadius.circular(6.0),
-                ),
-                focusedBorder: OutlineInputBorder(
-                  borderSide: BorderSide(color: Color(0xFF1E3984), width: 1.0),
-                ),
-              ),
-              keyboardType: TextInputType.emailAddress,
-              validator: (value) {
-                if (value == null || value.isEmpty) {
-                  return 'Por favor, ingrese su email';
-                }
-                return null;
-              },
-              onSaved: (value) {
-                _email = value!;
-              },
-            ),
+    return Container(
+      padding: const EdgeInsets.all(30),
+      child: Form(
+        key: _formKey,
+        child:
+            Column(crossAxisAlignment: CrossAxisAlignment.stretch, children: [
+          Image.asset(
+            'assets/logo_azul.png',
+            height: 208,
+            width: 211,
           ),
-          SizedBox(height: 7),
-          SizedBox(
-            width: 275.39,
-            height: 75,
-            child: TextFormField(
-              decoration: InputDecoration(
-                labelText: 'Contraseña admin',
-                filled: true,
-                fillColor: Color(0xFFDBE2F6),
-                enabledBorder: OutlineInputBorder(
-                  borderSide: BorderSide(color: Color(0xFF1E3984), width: 1.0),
-                ),
-                focusedBorder: OutlineInputBorder(
-                  borderSide: BorderSide(color: Color(0xFF1E3984), width: 1.0),
-                ),
-              ),
-              obscureText: true,
-              validator: (value) {
-                if (value == null || value.isEmpty) {
-                  return 'Por favor, ingrese su contraseña';
-                }
-                return null;
-              },
-              onSaved: (value) {
-                _password = value!;
-              },
-            ),
+          TextFormField(
+            controller: _email,
+            decoration: const InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: 'Ej: pepito@unmsm.edu.pe',
+                label: Text("Correo electrónico o telefono")),
           ),
-          SizedBox(height: 7),
-          SizedBox(
-            width: 150,
-            height: 48,
-            child: ElevatedButton(
-              onPressed: _submit,
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Color(0xFF1E3984),
-                foregroundColor: Color(0xFFF5F5F5),
-                textStyle: TextStyle(
-                  fontWeight: FontWeight.bold,
-                  fontSize: 16,
-                ),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(10.0),
-                ),
-              ),
-              child: Text('Ingresar'),
-            ),
+          const SizedBox(height: 10),
+          TextFormField(
+            controller: _password,
+            obscureText: true,
+            decoration: const InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: 'Ej: 123456',
+                label: Text("Contraseña")),
           ),
-        ],
+          const SizedBox(height: 15),
+          ElevatedButton(
+            onPressed: () {
+              _submit();
+            },
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Color(0xFF1E3984),
+              foregroundColor: Colors.white,
+              textStyle: TextStyle(
+                fontSize: 14,
+                fontWeight: FontWeight.bold,
+              ),
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(10.0),
+              ),
+              shadowColor: Colors.black.withAlpha(128),
+              elevation: 10,
+            ),
+            child: Text('Iniciar sesión'),
+          ),
+          const SizedBox(height: 10),
+          if (widget.index == 1) ...[
+            const Divider(),
+            const SizedBox(height: 10),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const Text('Aun no eres usuario?',
+                    style: TextStyle(fontSize: 14)),
+                TextButton(
+                  onPressed: null,
+                  child: const Text(
+                    'Registrate',
+                    style: TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(height: 10),
+            TextButton(
+              onPressed: null,
+              child: const Text('Olvidaste tu contraseña?'),
+            )
+          ]
+        ]),
       ),
     );
   }
