--- lib/features/features.dart
@@ -1,14 +1,0 @@
-// Auth Feature
-export 'auth/screens/login_screen.dart';
-export 'auth/widgets/login_form.dart';
-
-// Admin Feature
-export 'admin/screens/screens.dart';
-export 'admin/widgets/widgets.dart';
-export 'admin/main.dart';
-
-// Shared Components
-export 'shared/widgets/header.dart';
-export 'shared/widgets/footer.dart';
-export 'shared/screens/pick_user_screen.dart';
-export 'shared/screens/splash_screen.dart';
