+++ lib/features/shared/widgets/widgets.dart
@@ -1,0 +1,3 @@
+export 'footer.dart';
+export 'header.dart';
+export 'pick_user_card.dart';
